+++ Explorer/Assets/DCL/AuthenticationScreenFlow/Assets/VersionUpdatePopup.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 56fb9da1b08a38e4b8c8379394dd6aa6
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
