+++ Explorer/Assets/DCL/PluginSystem/Global/DynamicSettings.cs
@@ -1,3 +1,4 @@
+using DCL.AuthenticationScreenFlow;
 using DCL.Chat;
 using DCL.Input;
 using DCL.MapRenderer;
@@ -19,5 +20,6 @@ namespace DCL.PluginSystem.Global
         [field: SerializeField] public AssetReferenceT<CursorSettings> CursorSettings { get; private set; }
         [field: SerializeField] public AssetReferenceGameObject MainUIView { get; private set; }
         [field: SerializeField] public AssetReferenceT<AudioMixer> GeneralAudioMixer { get; private set; }
+        [field: SerializeField] public AssetReferenceT<LauncherRedirectionScreenView> AppVerRedirectionScreenPrefab { get; private set; }
     }
 }
